feat(FN-3796): code markups
