+++ libs/common/src/constants/index.ts
@@ -34,5 +34,6 @@ export * from './facility-day-count-basis';
 export * from './e2e-fixtures';
 export * from './regex';
 export * from './status';
+export * from './record-correction-status';
 
 export * as PAYLOAD_VERIFICATION from './payload-verification';
