+++ trade-finance-manager-ui/server/controllers/utilisation-reports/helpers/map-record-correction-status.test.ts
@@ -1,10 +1,10 @@
-import { setRecordCorrectionStatus } from './set-record-correction-status';
+import { mapToRecordCorrectionStatus } from './map-record-correction-status';
 
 describe('set-record=correction-status', () => {
-  describe('setRecordCorrectionStatus', () => {
+  describe('mapToRecordCorrectionStatus', () => {
     describe('isCompleted is true', () => {
       it('should return "RECORD_CORRECTION_RECEIVED" and "Record correction received"', () => {
-        const result = setRecordCorrectionStatus(true);
+        const result = mapToRecordCorrectionStatus(true);
 
         const expected = {
           displayStatus: 'Record correction received',
@@ -17,7 +17,7 @@ describe('set-record=correction-status', () => {
 
     describe('isCompleted is false', () => {
       it('should return "RECORD_CORRECTION_SENT" and "Record correction sent"', () => {
-        const result = setRecordCorrectionStatus(false);
+        const result = mapToRecordCorrectionStatus(false);
 
         const expected = {
           displayStatus: 'Record correction sent',
